Move BeatmapMetadataDisplay to its own class
