+++ osu.Game/Screens/Play/PlayerLoader.cs
@@ -2,7 +2,6 @@
 // See the LICENCE file in the repository root for full licence text.
 
 using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using osu.Framework.Allocation;
@@ -12,21 +11,15 @@ using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Sprites;
 using osu.Framework.Input;
-using osu.Framework.Localisation;
 using osu.Framework.Screens;
 using osu.Framework.Threading;
-using osu.Game.Beatmaps;
 using osu.Game.Configuration;
 using osu.Game.Graphics;
 using osu.Game.Graphics.Containers;
-using osu.Game.Graphics.Sprites;
-using osu.Game.Graphics.UserInterface;
 using osu.Game.Input;
 using osu.Game.Overlays;
 using osu.Game.Overlays.Notifications;
-using osu.Game.Rulesets.Mods;
 using osu.Game.Screens.Menu;
-using osu.Game.Screens.Play.HUD;
 using osu.Game.Screens.Play.PlayerSettings;
 using osu.Game.Users;
 using osuTK;
@@ -347,162 +340,6 @@ namespace osu.Game.Screens.Play
                 Background.BlurAmount.Value = BACKGROUND_BLUR;
 
                 BackgroundBrightnessReduction = true;
-            }
-        }
-
-        protected class BeatmapMetadataDisplay : Container
-        {
-            private class MetadataLine : Container
-            {
-                public MetadataLine(string left, string right)
-                {
-                    AutoSizeAxes = Axes.Both;
-                    Children = new Drawable[]
-                    {
-                        new OsuSpriteText
-                        {
-                            Anchor = Anchor.TopCentre,
-                            Origin = Anchor.TopRight,
-                            Margin = new MarginPadding { Right = 5 },
-                            Colour = OsuColour.Gray(0.8f),
-                            Text = left,
-                        },
-                        new OsuSpriteText
-                        {
-                            Anchor = Anchor.TopCentre,
-                            Origin = Anchor.TopLeft,
-                            Margin = new MarginPadding { Left = 5 },
-                            Text = string.IsNullOrEmpty(right) ? @"-" : right,
-                        }
-                    };
-                }
-            }
-
-            private readonly WorkingBeatmap beatmap;
-            private readonly Bindable<IReadOnlyList<Mod>> mods;
-            private readonly Drawable facade;
-            private LoadingAnimation loading;
-            private Sprite backgroundSprite;
-
-            public IBindable<IReadOnlyList<Mod>> Mods => mods;
-
-            public bool Loading
-            {
-                set
-                {
-                    if (value)
-                    {
-                        loading.Show();
-                        backgroundSprite.FadeColour(OsuColour.Gray(0.5f), 400, Easing.OutQuint);
-                    }
-                    else
-                    {
-                        loading.Hide();
-                        backgroundSprite.FadeColour(Color4.White, 400, Easing.OutQuint);
-                    }
-                }
-            }
-
-            public BeatmapMetadataDisplay(WorkingBeatmap beatmap, Bindable<IReadOnlyList<Mod>> mods, Drawable facade)
-            {
-                this.beatmap = beatmap;
-                this.facade = facade;
-
-                this.mods = new Bindable<IReadOnlyList<Mod>>();
-                this.mods.BindTo(mods);
-            }
-
-            [BackgroundDependencyLoader]
-            private void load()
-            {
-                var metadata = beatmap.BeatmapInfo?.Metadata ?? new BeatmapMetadata();
-
-                AutoSizeAxes = Axes.Both;
-                Children = new Drawable[]
-                {
-                    new FillFlowContainer
-                    {
-                        AutoSizeAxes = Axes.Both,
-                        Origin = Anchor.TopCentre,
-                        Anchor = Anchor.TopCentre,
-                        Direction = FillDirection.Vertical,
-                        Children = new[]
-                        {
-                            facade.With(d =>
-                            {
-                                d.Anchor = Anchor.TopCentre;
-                                d.Origin = Anchor.TopCentre;
-                            }),
-                            new OsuSpriteText
-                            {
-                                Text = new LocalisedString((metadata.TitleUnicode, metadata.Title)),
-                                Font = OsuFont.GetFont(size: 36, italics: true),
-                                Origin = Anchor.TopCentre,
-                                Anchor = Anchor.TopCentre,
-                                Margin = new MarginPadding { Top = 15 },
-                            },
-                            new OsuSpriteText
-                            {
-                                Text = new LocalisedString((metadata.ArtistUnicode, metadata.Artist)),
-                                Font = OsuFont.GetFont(size: 26, italics: true),
-                                Origin = Anchor.TopCentre,
-                                Anchor = Anchor.TopCentre,
-                            },
-                            new Container
-                            {
-                                Size = new Vector2(300, 60),
-                                Margin = new MarginPadding(10),
-                                Origin = Anchor.TopCentre,
-                                Anchor = Anchor.TopCentre,
-                                CornerRadius = 10,
-                                Masking = true,
-                                Children = new Drawable[]
-                                {
-                                    backgroundSprite = new Sprite
-                                    {
-                                        RelativeSizeAxes = Axes.Both,
-                                        Texture = beatmap?.Background,
-                                        Origin = Anchor.Centre,
-                                        Anchor = Anchor.Centre,
-                                        FillMode = FillMode.Fill,
-                                    },
-                                    loading = new LoadingAnimation { Scale = new Vector2(1.3f) }
-                                }
-                            },
-                            new OsuSpriteText
-                            {
-                                Text = beatmap?.BeatmapInfo?.Version,
-                                Font = OsuFont.GetFont(size: 26, italics: true),
-                                Origin = Anchor.TopCentre,
-                                Anchor = Anchor.TopCentre,
-                                Margin = new MarginPadding
-                                {
-                                    Bottom = 40
-                                },
-                            },
-                            new MetadataLine("Source", metadata.Source)
-                            {
-                                Origin = Anchor.TopCentre,
-                                Anchor = Anchor.TopCentre,
-                            },
-                            new MetadataLine("Mapper", metadata.AuthorString)
-                            {
-                                Origin = Anchor.TopCentre,
-                                Anchor = Anchor.TopCentre,
-                            },
-                            new ModDisplay
-                            {
-                                Anchor = Anchor.TopCentre,
-                                Origin = Anchor.TopCentre,
-                                AutoSizeAxes = Axes.Both,
-                                Margin = new MarginPadding { Top = 20 },
-                                Current = mods
-                            }
-                        },
-                    }
-                };
-
-                Loading = true;
             }
         }
